Move LEB128 stuff to its own file
Also rename some of the functions.
